no more int long etc

+++ tools/test-tools/sam-analyze/database.hpp
@@ -14,7 +14,7 @@ class PREP_STM {
     private :
         sqlite3 *db;
         sqlite3_stmt * stmt;
-        int status;
+        int16_t status;
 
     public :
         typedef std::vector< std::string > str_vec;
@@ -29,30 +29,30 @@ class PREP_STM {
             sqlite3_finalize( stmt );
         }
 
-        bool ok_or_done( int st ) {
+        bool ok_or_done( int16_t st ) {
             return ( SQLITE_OK == st || SQLITE_DONE == st );
         }
 
-        int get_status( void ) const { return status; }
+        int16_t get_status( void ) const { return status; }
         
-        void print_status( int status ) {
+        void print_status( int16_t status ) {
             const char * e = sqlite3_errstr( status );
             std::cout << "status:" << e << std::endl;
         }
         
-        int step( void ) {
+        int16_t step( void ) {
             if ( ok_or_done( status ) ) { status = sqlite3_step( stmt ); }
             return status;
         }
 
-        int reset( void ) {
+        int16_t reset( void ) {
             if ( ok_or_done( status ) ) { status = sqlite3_reset( stmt ); }
             return status;
         }
 
-        int bind_str_vec( const str_vec& data ) {
+        int16_t bind_str_vec( const str_vec& data ) {
             if ( ok_or_done( status ) ) {
-                int idx = 1;
+                int16_t idx = 1;
                 for ( auto it = data.begin(); ok_or_done( status ) && it != data.end(); it++ ) {
                     status = sqlite3_bind_text( stmt,
                                                 idx++,
@@ -63,21 +63,21 @@ class PREP_STM {
             return status;
         }
 
-        int bind_str( const std::string& data ) {
+        int16_t bind_str( const std::string& data ) {
             if ( ok_or_done( status ) && !data.empty() ) {
                 status = sqlite3_bind_text( stmt, 1, data.c_str(), data.length(), SQLITE_STATIC );
             }
             return status;
         }
         
-        int bind_and_step( const str_vec& data ) {
+        int16_t bind_and_step( const str_vec& data ) {
             bind_str_vec( data );
             step();
             reset();
             return status;
         }
 
-        int read_text( std::string& result ) {
+        int16_t read_text( std::string& result ) {
             if ( SQLITE_ROW == step() ) {
                 const unsigned char * txt = sqlite3_column_text( stmt, 0 );
                 std::string s( (const char *)txt );
@@ -88,17 +88,17 @@ class PREP_STM {
             return status;
         }
 
-        int read_text_1( const std::string& data, std::string& result ) {
+        int16_t read_text_1( const std::string& data, std::string& result ) {
             bind_str( data );
             return read_text( result );
         }
         
-        int read_text_n( const str_vec& data, std::string& result ) {
+        int16_t read_text_n( const str_vec& data, std::string& result ) {
             bind_str_vec( data );
             return read_text( result );            
         }
 
-        int read_long( unsigned long& result ) {
+        int16_t read_long( unsigned long& result ) {
             step();
             if ( SQLITE_ROW == status ) {
                 result = sqlite3_column_int64( stmt, 0 );
@@ -110,23 +110,23 @@ class PREP_STM {
             return status;
         }
 
-        int read_long_1( const std::string& data, unsigned long& result ) {
+        int16_t read_long_1( const std::string& data, unsigned long& result ) {
             bind_str( data );
             return read_long( result );
         }
         
-        int read_long_n( const str_vec& data, unsigned long& result ) {
+        int16_t read_long_n( const str_vec& data, unsigned long& result ) {
             bind_str_vec( data );
             return read_long( result );            
         }
 
-        int read_row( str_vec& data, int count ) {
+        int16_t read_row( str_vec& data, int16_t count ) {
             data . clear();
             if ( status == SQLITE_OK || status == SQLITE_ROW ) {
                 status = sqlite3_step( stmt );
             }
             if ( SQLITE_ROW == status ) {
-                for ( int i = 0; i < count; ++ i ) {
+                for ( int16_t i = 0; i < count; ++ i ) {
                     const unsigned char * txt = sqlite3_column_text( stmt, i );
                     data . push_back( std::string( ( const char * )txt ) );
                 }
@@ -134,7 +134,7 @@ class PREP_STM {
             return status;
         }
         
-        bool read_2_longs( unsigned long& v1, unsigned long& v2 ) {
+        bool read_2_longs( uint64_t& v1, uint64_t& v2 ) {
             if ( status == SQLITE_OK || status == SQLITE_ROW ) {
                 status = sqlite3_step( stmt );
             }
@@ -156,7 +156,7 @@ class DB {
         typedef std::vector< std::string > str_vec;
 
         sqlite3 *db;
-        int rc_open;
+        int16_t rc_open;
         
         DB( const char * name ) : db( nullptr ) {
             rc_open = sqlite3_open( name, &db );
@@ -169,8 +169,8 @@ class DB {
         }
 
         
-        int exec( const char * stm ) {
-            int res = -1;
+        int16_t exec( const char * stm ) {
+            int16_t res = -1;
             if ( SQLITE_OK == rc_open && nullptr != db ) {
                 char * errMsg;
                 res = sqlite3_exec( db, stm, nullptr, 0, &errMsg );
@@ -183,21 +183,21 @@ class DB {
 
         sqlite3* get_db_handle( void ) const { return db; }
 
-        int begin_transaction( void ) { return exec( "BEGIN TRANSACTION" ); }
-        int commit_transaction( void ) { return exec( "COMMIT TRANSACTION" ); }
+        int16_t begin_transaction( void ) { return exec( "BEGIN TRANSACTION" ); }
+        int16_t commit_transaction( void ) { return exec( "COMMIT TRANSACTION" ); }
         
-        int clear_table( const std::string& tbl_name, bool vacuum = false ) {
+        int16_t clear_table( const std::string& tbl_name, bool vacuum = false ) {
             std::string sql = "DELETE FROM " + tbl_name + ";";
             PREP_STM drop( db, sql );
-            int res = drop.step();
+            int16_t res = drop . step();
             if ( ok_or_done( res ) && vacuum ) {
                 res = exec( "VACUUM" );
             }
             return res;
         }
         
-        int clear_tables( const str_vec& tables, bool vacuum = true ) {
-            int res = SQLITE_OK;
+        int16_t clear_tables( const str_vec& tables, bool vacuum = true ) {
+            int16_t res = SQLITE_OK;
             for ( auto it = tables.begin(); ok_or_done( res ) && it != tables.end(); it++ ) {
                 res = clear_table( *it );
             }
@@ -207,11 +207,11 @@ class DB {
             return res;
         }
         
-        bool ok_or_done( int status ) {
+        bool ok_or_done( int16_t status ) {
             return ( SQLITE_OK == status || SQLITE_DONE == status );
         }
 
-        void print_status( int status ) {
+        void print_status( int16_t status ) {
             const char * e = sqlite3_errstr( status );
             std::cout << "status:" << e << std::endl;
         }
